+++ src/main/java/com/project/finalproject/login/AuthController.java
@@ -1,9 +1,9 @@
-package com.project.finalproject.signup.controller;
+package com.project.finalproject.login;
 
-import com.project.finalproject.signup.dto.ApplicantSignupReqDTO;
-import com.project.finalproject.signup.dto.CompanySignupReqDTO;
-import com.project.finalproject.signup.service.ApplicantSignupService;
-import com.project.finalproject.signup.service.CompanySignupService;
+import com.project.finalproject.login.dto.ApplicantSignupReqDTO;
+import com.project.finalproject.login.dto.CompanySignupReqDTO;
+import com.project.finalproject.login.signup.applicant.ApplicantSignupService;
+import com.project.finalproject.login.signup.company.CompanySignupService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class SignupController {
+public class AuthController {
 
     private final ApplicantSignupService applicantSignupService;
     private final CompanySignupService companySignupService;
