+++ src/main/java/com/project/finalproject/login/login/applicant/ApplicantLoginRepository.java
@@ -1,0 +1,4 @@
+package com.project.finalproject.login.login.applicant;
+
+public interface ApplicantLoginRepository {
+}
