+++ src/main/java/com/project/finalproject/login/login/applicant/ApplicantLoginService.java
@@ -1,0 +1,4 @@
+package com.project.finalproject.login.login.applicant;
+
+public class ApplicantLoginService {
+}
