+++ src/main/java/com/project/finalproject/login/login/company/CompanyLoginRepository.java
@@ -1,0 +1,4 @@
+package com.project.finalproject.login.login.company;
+
+public interface CompanyLoginRepository {
+}
