+++ src/main/java/com/project/finalproject/login/login/company/CompanyLoginService.java
@@ -1,0 +1,4 @@
+package com.project.finalproject.login.login.company;
+
+public class CompanyLoginService {
+}
