+++ src/main/java/com/project/finalproject/login/signup/applicant/ApplicantSignupRepository.java
@@ -1,4 +1,4 @@
-package com.project.finalproject.signup.repository;
+package com.project.finalproject.login.signup.applicant;
 
 import com.project.finalproject.applicant.entity.Applicant;
 import org.springframework.data.jpa.repository.JpaRepository;
