+++ src/main/java/com/project/finalproject/login/signup/applicant/ApplicantSignupService.java
@@ -1,9 +1,8 @@
-package com.project.finalproject.signup.service;
+package com.project.finalproject.login.signup.applicant;
 
-import com.project.finalproject.company.entity.Company;
+import com.project.finalproject.applicant.entity.Applicant;
 import com.project.finalproject.global.jwt.utils.JwtUtil;
-import com.project.finalproject.signup.dto.CompanySignupReqDTO;
-import com.project.finalproject.signup.repository.CompanySignupRepository;
+import com.project.finalproject.login.dto.ApplicantSignupReqDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -13,26 +12,25 @@ import javax.transaction.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class CompanySignupService {
+public class ApplicantSignupService {
 
-    private final CompanySignupRepository companySignupRepository;
+    private final ApplicantSignupRepository applicantSignupRepository;
     private final JwtUtil jwtUtil;
-
     private final PasswordEncoder passwordEncoder;
 
     /**
      * 회원가입
      */
-    public String signUp(CompanySignupReqDTO reqDTO){
-        if(companySignupRepository.existsCompanyByEmail(reqDTO.getEmail())){
+    public String signUp(ApplicantSignupReqDTO reqDTO){
+        if(applicantSignupRepository.existsApplicantByEmail(reqDTO.getEmail())){
             return reqDTO.getEmail() + "는 이미 존재하는 아이디 입니다.";
         }
 
         reqDTO.setPassword(passwordEncoding(reqDTO.getPassword()));
 
-        Company company = reqDTO.toEntity();
+        Applicant applicant = reqDTO.toEntity();
 
-        companySignupRepository.save(company);
+        applicantSignupRepository.save(applicant);
 
         return "회원가입이 완료되었습니다.";
     }
@@ -43,5 +41,4 @@ public class CompanySignupService {
     private String passwordEncoding(String password){
         return passwordEncoder.encode(password);
     }
-
 }
