+++ src/main/java/com/project/finalproject/login/signup/company/CompanySignupService.java
@@ -1,9 +1,8 @@
-package com.project.finalproject.signup.service;
+package com.project.finalproject.login.signup.company;
 
-import com.project.finalproject.applicant.entity.Applicant;
+import com.project.finalproject.company.entity.Company;
 import com.project.finalproject.global.jwt.utils.JwtUtil;
-import com.project.finalproject.signup.dto.ApplicantSignupReqDTO;
-import com.project.finalproject.signup.repository.ApplicantSignupRepository;
+import com.project.finalproject.login.dto.CompanySignupReqDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -13,25 +12,26 @@ import javax.transaction.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class ApplicantSignupService {
+public class CompanySignupService {
 
-    private final ApplicantSignupRepository applicantSignupRepository;
+    private final CompanySignupRepository companySignupRepository;
     private final JwtUtil jwtUtil;
+
     private final PasswordEncoder passwordEncoder;
 
     /**
      * 회원가입
      */
-    public String signUp(ApplicantSignupReqDTO reqDTO){
-        if(applicantSignupRepository.existsApplicantByEmail(reqDTO.getEmail())){
+    public String signUp(CompanySignupReqDTO reqDTO){
+        if(companySignupRepository.existsCompanyByEmail(reqDTO.getEmail())){
             return reqDTO.getEmail() + "는 이미 존재하는 아이디 입니다.";
         }
 
         reqDTO.setPassword(passwordEncoding(reqDTO.getPassword()));
 
-        Applicant applicant = reqDTO.toEntity();
+        Company company = reqDTO.toEntity();
 
-        applicantSignupRepository.save(applicant);
+        companySignupRepository.save(company);
 
         return "회원가입이 완료되었습니다.";
     }
@@ -42,4 +42,5 @@ public class ApplicantSignupService {
     private String passwordEncoding(String password){
         return passwordEncoder.encode(password);
     }
+
 }
